requirements.txt
Django==2.1.5
django_environ==0.4.5
djangorestframework_jwt==1.11.0
environ==1.0
astroid==2.1.0
awsebcli==3.14.6
botocore==1.12.15
cement==2.8.2
colorama==0.3.9
docutils==0.14
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
pathspec==0.5.5
psycopg2==2.7.6.1
pylint==2.2.2
python-dateutil==2.7.3
pytz==2018.9
PyYAML==3.13
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
urllib3==1.22
wrapt==1.11.0